requirements.txt
asgiref==3.8.1
attrs==25.4.0
beautifulsoup4==4.14.3
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.1
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
et_xmlfile==2.0.0
greenlet==3.3.1
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
numpy==2.3.4
openpyxl==3.1.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.58.0
ply==3.11
psycopg2==2.9.11
puresnmp==2.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycryptodomex==3.23.0
pyee==13.0.1
PyJWT==2.10.1
pysmi==1.6.2
pysnmp==7.1.22
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
reportlab==4.4.4
requests==2.32.5
rpds-py==0.30.0
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.3
t61codec==1.0.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
x690==1.0.0.post1
django-auditlog==3.4.1
